feat: Pass AbortSignal from getTileData down to geotiff read fn (#69)
When a user pans very quickly, deck.gl may make many requests for tiles.
deck.gl will cancel the requests for tiles outside the current viewport
that haven't finished. We should pass the signal on to the actual
request being made in `geotiff.js` so that it will in turn cancel the
request for us.

See upstream docs:
https://deck.gl/docs/api-reference/geo-layers/tile-layer#gettiledata